[feat] edit-hotel

+++ src/app/component/portal/pages/hotel/edit-hotel/edit-hotel.component.ts
@@ -27,7 +27,7 @@ export class EditHotelComponent implements OnInit {
   isUpdate: boolean;
   titleFormHotel = '';
   fileList: NzUploadFile[] = [];
-  listURLFiles: any[] = [];
+  hotelFileIds: any[] = [];
   uploadUrl = '';
   listUtilityHotel = [];
   uploadHeader: any;
@@ -202,7 +202,7 @@ export class EditHotelComponent implements OnInit {
       isActive: true
     });
     this.fileList = [];
-    this.listURLFiles = [];
+    this.hotelFileIds = [];
     this.uploadUrl = `${this.configService.getConfig().api.baseUrl}/Upload/UploadHotelImage?hotelId=0`;
   }
 
@@ -319,8 +319,6 @@ export class EditHotelComponent implements OnInit {
   }
 
   handleChangeImages({ file, fileList }: NzUploadChangeParam, form: any): void {
-    console.log('file.response: ', file);
-
     const status = file.status;
     if (status === 'done') {
       this.msg.success(`file ${file.name} tải lên thành công.`);
@@ -328,7 +326,7 @@ export class EditHotelComponent implements OnInit {
       if (form === 'hotel') {
         setTimeout(() => {
           if (this.fileList.length > 0) {
-            this.fileList[this.fileList.length - 1].uid = file.response.hotelFileId.toString();
+            this.fileList[this.fileList.length - 1].hotelFileId = file.response.hotelFileId.toString();
             this.fileList[this.fileList.length - 1].url = `${this.configService.getConfig().api.baseUrl}/${file.response.path}`;
             // this.fileList[this.fileList.length - 1] = {
             //   uid: file.response.hotelFileId.toString(),
@@ -340,8 +338,8 @@ export class EditHotelComponent implements OnInit {
             // };
           }
         }, 200);
-        this.listURLFiles.push(file.response.hotelFileId);
-        this.formAddHotel.controls['hotelFileIds'].setValue(this.listURLFiles);
+        this.hotelFileIds.push(file.response.hotelFileId);
+        this.formAddHotel.controls['hotelFileIds'].setValue(this.hotelFileIds);
       } else if (form === 'room') {
         // setTimeout(() => {
         //   if (this.fileList.length > 0) {
@@ -359,34 +357,36 @@ export class EditHotelComponent implements OnInit {
 
 
   handleRemoveImageHotel = async (file: NzUploadFile): Promise<void> => {
-    const idHotelImage = file.uid;
-    this.generalService.deleteHotelImageByID(idHotelImage).subscribe(
-      {
-        next: (res) => {
-          if (res) {
-            if (res.ret && res.ret.length > 0) {
-              res.ret.forEach((el: any) => {
-                if (el.code === 0) {
-                  this.msg.success(`Đã xoá ảnh ${file.name}.`);
-                  // this.getListData();
-                } else if (res.code === 404) {
-                  this.msg.error(`Không tìm thấy ảnh ${file.name}.`);
-                } else {
-                  this.msg.error(`Đã có lỗi xảy ra. Không thể xoá ảnh ${file.name}`);
-                }
-              });
+    const idHotelImage = file.hotelFileId;
+    if (idHotelImage) {
+      this.generalService.deleteHotelImageByID(idHotelImage).subscribe(
+        {
+          next: (res) => {
+            if (res) {
+              if (res.ret && res.ret.length > 0) {
+                res.ret.forEach((el: any) => {
+                  if (el.code === 0) {
+                    this.msg.success(`Đã xoá ảnh ${file.name}.`);
+                    // this.getListData();
+                  } else if (res.code === 404) {
+                    this.msg.error(`Không tìm thấy ảnh ${file.name}.`);
+                  } else {
+                    this.msg.error(`Đã có lỗi xảy ra. Không thể xoá ảnh ${file.name}`);
+                  }
+                });
+              }
+            } else {
+              this.notificationService.showNotification(Constant.ERROR, `Hệ thống gặp lỗi, xoá ảnh ${file.name} thật bại.`);
             }
-          } else {
+          },
+          error: (error) => {
             this.notificationService.showNotification(Constant.ERROR, `Hệ thống gặp lỗi, xoá ảnh ${file.name} thật bại.`);
-          }
-        },
-        error: (error) => {
-          this.notificationService.showNotification(Constant.ERROR, `Hệ thống gặp lỗi, xoá ảnh ${file.name} thật bại.`);
-        },
-        complete: () => {
-        },
+          },
+          complete: () => {
+          },
 
-      }
-    );
+        }
+      );
+    }
   };
 }
